lab-session4-2609/workflow2-stepfunctions/fetch-image/src/requirements.txt
boto3==1.34.144
Pillow==10.0.0
